discreteLogPohligHellman optimizations.
Fixed the original "raw" implementation of the Pohlig-Hellman algorithm which was correct but not very efficient with computing babyGiant bounds (by recalculating the value of sqrt(p) in every iteration) and re-evaluating a few booleans. Instead changed it to minimally use resources and compute values as they are needed and not a second sooner.