MathUtil improvements.
Optimized calculations and checks in discreteLogLinearSearchFixedInput and discreteLogBabyGiantFixedInput to remove unnecessary operations and further improve the performance. Made sure that every operation performed in the aforementioned discreteLog methods is actually required by rewriting parts of the code to not calculate the last value when we don't know if the loop is going to iterate one more time or not and also significantly modifying the order check logic in babyGiant generateBoth case to be extremely conservative with calculations.

Fixed some comments.